fix: remove circular dependency between base.ts and command_types.ts for Opts
