+++ src/commands/index.ts
@@ -9,8 +9,8 @@ import {NodeCommand} from './node/index.js';
 import {RelayCommand} from './relay.js';
 import {AccountCommand} from './account.js';
 import {DeploymentCommand} from './deployment.js';
-import {type Opts} from '../types/command_types.js';
 import {ExplorerCommand} from './explorer.js';
+import {type Opts} from './base.js';
 
 /**
  * Return a list of Yargs command builder to be exposed through CLI
